Abolish fetch_history, work on fetching history data
We cannot rely on the history API to provide all the data we might need from the primary API, so we have to revise our approach. While we can now fetch from either API (and cache what we can get of the primary data from the history API), we do not yet have the ability to fetch history data from the object.